# expression: handle max_allowed_packet warnings for repeat function.